Fix another compilation error in QDateTime
Similar fix as 538e9fa5689a1afe0d821b84cdfeaa1656973083 : gcc 7.5,
used e.g. on Jetson, is not able to resolve the add_overflow overload
without help.

+++ src/corelib/time/qdatetime.cpp
@@ -2663,7 +2663,9 @@ bool QDateTimePrivate::epochMSecsToLocalTime(qint64 msecs, QDate *localDate, QTi
                 : QDateTimePrivate::StandardTime;
         }
 
-        if (add_overflow(msecs, sys.d->offsetFromUtc(msecs) * MSECS_PER_SEC, &msecs))
+        // NB: cast to qint64 here is important to make sure a matching
+        // add_overflow is found, GCC 7.5.0 fails without this cast
+        if (add_overflow(msecs, qint64(sys.d->offsetFromUtc(msecs) * MSECS_PER_SEC), &msecs))
             return false;
         msecsToTime(msecs, localDate, localTime);
         return true;
